rm MarketPlace panels and extra SideNavigation component | updated GET_USER query to return coach_resume_reviews and seeker_resume_reviews
