+++ src/ResumeQ/components/Marketplace/Marketplace.js
@@ -2,11 +2,13 @@ import React from 'react';
 
 // sub imports
 import ReviewerList from './ReviewerList'
+import SideNavigation from './SideNavigation'
 
 const Marketplace = () => {
 
   return (
     <>
+      <SideNavigation />
       <ReviewerList />
     </>
   )
