--- src/ResumeQ/components/Marketplace/ReviewerList/subs/2_Panels/index.js
@@ -1,3 +1,0 @@
-import MarketplacePanels from './MarketplacePanels'
-
-export default MarketplacePanels
