+++ src/ResumeQ/components/Marketplace/SideNavigation/SideNavigation.scss
@@ -1,4 +1,4 @@
-@import '../../../global/styles/variables';
+@import '../../../../global/styles/variables';
 
 .RQNav {
   z-index: 900;
