+++ src/ResumeQ/components/Marketplace/SideNavigation/index.js
@@ -1,0 +1,4 @@
+import SideNavigation from './SideNavigation';
+
+
+export default SideNavigation;
