--- src/ResumeQ/components/SideNavigation/index.js
@@ -1,3 +1,0 @@
-import SideNaviation from './SideNavigation'
-
-export default SideNaviation;
